+++ src/main/java/com/example/yeobee/core/auth/application/AppleAuthService.java
@@ -7,8 +7,8 @@ import com.example.yeobee.core.auth.domain.AuthProviderType;
 import com.example.yeobee.core.auth.dto.request.AppleLoginRequestDto;
 import com.example.yeobee.core.auth.dto.response.AppleAuthTokenResponseDto;
 import com.example.yeobee.core.auth.dto.response.TokenResponseDto;
-import com.example.yeobee.core.user.entity.User;
-import com.example.yeobee.core.user.repository.UserRepository;
+import com.example.yeobee.core.user.domain.User;
+import com.example.yeobee.core.user.domain.UserRepository;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import java.io.Reader;
