+++ src/main/java/com/example/yeobee/core/auth/application/AuthService.java
@@ -4,8 +4,8 @@ import com.example.yeobee.common.exception.BusinessException;
 import com.example.yeobee.common.exception.ErrorCode;
 import com.example.yeobee.core.auth.domain.*;
 import com.example.yeobee.core.auth.dto.response.TokenResponseDto;
-import com.example.yeobee.core.user.entity.User;
-import com.example.yeobee.core.user.repository.UserRepository;
+import com.example.yeobee.core.user.domain.User;
+import com.example.yeobee.core.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
