+++ src/main/java/com/example/yeobee/core/user/domain/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.yeobee.core.user.repository;
+package com.example.yeobee.core.user.domain;
 
-import com.example.yeobee.core.user.entity.User;
+import com.example.yeobee.core.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
